voicelegacy-backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
aiosqlite
pydantic-settings
httpx
prometheus-client
passlib[bcrypt]
python-jose
python-multipart
websockets
numpy
chromadb
sentence-transformers
transformers
boto3
minio
websockets==10.4

asyncpg

aiofiles
email-validator==2.2.0
email-validator==2.2.0
asyncpg==0.29.0
python-multipart==0.0.9
pydantic-settings==2.4.0

